Began work on nested variables
