+++ packages/scenes-app/src/demos/index.ts
@@ -6,7 +6,6 @@ import { getFlexLayoutTest } from './flexLayout';
 import { getGridLayoutTest } from './grid';
 import { getGridWithRowLayoutTest } from './gridWithRow';
 import { getLazyLoadDemo } from './lazyLoad';
-import { getNestedScene } from './nestedScene';
 import { getPanelContextDemoScene } from './panelContext';
 import { getPanelMenuTest } from './panelMenu';
 import { getPanelRepeaterTest } from './panelRepeater';
@@ -26,6 +25,7 @@ import { getAdhocFiltersDemo } from './adhocFiltersDemo';
 import { getTransformationsTest } from './transformations';
 import { getDynamicVizOptionsTest } from './dynamicPanelOptions';
 import { getDataFilteringTest } from './filteringData';
+import { getNestedScenesAndVariablesDemo } from './nestedVariables';
 
 export interface DemoDescriptor {
   title: string;
@@ -44,7 +44,6 @@ export function getDemos(): DemoDescriptor[] {
     { title: 'Lazy load', getPage: getLazyLoadDemo },
     { title: 'Variables', getPage: getVariablesDemo },
     { title: 'Adhoc filters', getPage: getAdhocFiltersDemo },
-    { title: 'Nested scene', getPage: getNestedScene },
     { title: 'With drilldowns', getPage: getDrilldownsAppPageScene },
     { title: 'Query editor', getPage: getQueryEditorDemo },
     { title: 'Dynamic page', getPage: getDynamicPageDemo },
@@ -61,5 +60,6 @@ export function getDemos(): DemoDescriptor[] {
     { title: 'Dynamic panel options and field config', getPage: getDynamicVizOptionsTest },
     { title: 'Data filtering', getPage: getDataFilteringTest },
     { title: 'Docs examples', getPage: getDocsExamples },
+    { title: 'Nested scenes and variables', getPage: getNestedScenesAndVariablesDemo },
   ].sort((a, b) => a.title.localeCompare(b.title));
 }
